Created QRCodesExtractor.
Issue #121
