+++ GiniVision/Classes/GiniQRCodeDocument.swift
@@ -15,7 +15,8 @@ import Foundation
     }()
     public var isReviewable: Bool = false
     public var isImported: Bool = false
-    public lazy var extractedParameters: [String: String] = self.extractParameters(from: self.scannedString)
+    public lazy var extractedParameters: [String: String] = QRCodesExtractor
+        .extractParameters(from: self.scannedString, withFormat: self.qrCodeFormat)
     
     fileprivate let scannedString: String
     fileprivate let epc06912LinesCount = 12
@@ -32,11 +33,6 @@ import Foundation
             return nil
         }
     }()
-    
-    fileprivate enum QRCodesFormat {
-        case epc06912
-        case bezahlcode
-    }
     
     public init(scannedString: String) {
         self.data = scannedString.data(using: String.Encoding.isoLatin1) ?? Data(count: 0)
@@ -59,91 +55,5 @@ extension GiniQRCodeDocument {
             return self.scannedString == object.scannedString
         }
         return false
-    }
-}
-
-// MARK: Extractions
-
-extension GiniQRCodeDocument {
-    fileprivate func extractParameters(from string: String) -> [String: String] {
-        switch qrCodeFormat {
-        case .some(.bezahlcode):
-            return extractParameters(fromBezhalCodeString: string)
-        case .some(.epc06912):
-            return extractParameters(fromEPC06912CodeString: string)
-        case .none:
-            return [:]
-        }
-    }
-    
-    fileprivate func extractParameters(fromBezhalCodeString string: String) -> [String: String] {
-        var parameters: [String: String] = [:]
-        
-        if let queryParameters = URL(string: string)?.queryParameters {
-            
-            if let bic = queryParameters["bic"] as? String {
-                parameters["bic"] = bic
-            }
-            if let paymentRecipient = queryParameters["name"] as? String {
-                parameters["paymentRecipient"] = paymentRecipient
-            }
-            if let iban = queryParameters["iban"] as? String,
-                IBANValidator().isValid(iban: iban) {
-                parameters["iban"] = iban
-            }
-            if let paymentReference = queryParameters["reason"] as? String ??
-                queryParameters["reason1"] as? String {
-                parameters["paymentReference"] = paymentReference
-            }
-            if let amount = queryParameters["amount"] as? String,
-                let amountNormalized = normalize(amount: amount,
-                                                 currency: queryParameters["currency"] as? String ?? "EUR") {
-                parameters["amountToPay"] = amountNormalized
-            }
-        }
-        
-        return parameters
-    }
-    
-    fileprivate func extractParameters(fromEPC06912CodeString string: String) -> [String: String] {
-        let lines = string.splitlines
-        var parameters: [String: String] = [:]
-        
-        if !lines[4].isEmpty {
-            parameters["bic"] = lines[4]
-        }
-        
-        if !lines[5].isEmpty {
-            parameters["paymentRecipient"] = lines[5]
-        }
-        
-        if !lines[9].isEmpty {
-            parameters["paymentReference"] = lines[9]
-        }
-        
-        if IBANValidator().isValid(iban: lines[6]) {
-            parameters["iban"] = lines[6]
-        }
-        
-        if let amountToPay = normalize(amount: lines[7], currency: nil) {
-            parameters["amountToPay"] = amountToPay
-        }
-        
-        return parameters
-    }
-    
-    fileprivate func normalize(amount: String, currency: String?) -> String? {
-        let regexCurrency = try? NSRegularExpression(pattern: "[aA-zZ]", options: [])
-        let length = amount.count < 3 ? amount.count : 3
-        
-        if regexCurrency?.numberOfMatches(in: amount, options: [], range: NSRange(location: 0, length: length)) == 3 {
-            let currency = amount.substring(to: String.Index(encodedOffset: 3))
-            let quantity = amount.substring(from: String.Index(encodedOffset: 3))
-            return quantity + ":" + currency
-        } else if let currency = currency {
-            return amount + ":" + currency
-        }
-        
-        return nil
     }
 }
