+++ GiniVision/Classes/GiniVisionUtils.swift
@@ -68,7 +68,6 @@ internal func UIFontPreferred(_ weight: FontWeight, andSize size: CGFloat) -> UI
     }
 }
 
-
 internal struct AnimationDuration {
     static var slow = 1.0
     static var medium = 0.5
